move missing MDI icons to assets
