+++ src/PublicShareSidebarTrigger.vue
@@ -33,7 +33,7 @@
 <script>
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
 
-import MenuPeople from './components/missingMaterialDesignIcons/MenuPeople.vue'
+import MenuPeople from './assets/missingMaterialDesignIcons/MenuPeople.vue'
 
 export default {
