+++ src/assets/missingMaterialDesignIcons/MenuPeople.vue
